🚚 rename: AccountDeletionForm을 index로 변경 #42
